Converted long ids to ints

+++ Assets/Scripts/Data/GameData.cs
@@ -9,7 +9,7 @@ public class GameData {
         return Game.Inst.data;
     }
 
-    public long nextId = 1;
+    public int nextId = 1;
     public uint rngState = 0; // Will be set by Rng class.
 
     public GameOptions gameOptions = new();
@@ -18,7 +18,7 @@ public class GameData {
     public JobData jobData = new();
     public CameraData cameraData = new();
 
-    public long GetNextId() {
+    public int GetNextId() {
         return nextId++;
     }
